# swagger/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/MALTHEKUSK/ygo-api/1.0.0"
  } ],
  "info" : {
    "version" : "1.0.0",
    "title" : "YGO API",
    "description" : "YGO API"
  },
  "paths" : {
...
    "/events/recent/{limit}" : {
      "get" : {
        "summary" : "Get most recent events limited by parameter",
        "description" : "Returns an array of recent events and their related information limited to the amount of the parameter.",
        "tags" : [ "events" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "limit",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Array of event information",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/events/type/{slug}" : {
      "get" : {
        "summary" : "Get all events of a type based on slug.",
        "description" : "Returns an array of events of a type and their related information.",
        "tags" : [ "events" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "slug",
          "schema" : {
            "type" : "string",
            "enum" : [ "wcq", "ycs", "team-ycs", "remote-ycs" ]
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Array of event information",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/events/{slug}" : {
      "get" : {
        "summary" : "Get event by slug",
        "description" : "Returns event information based on slug.",
        "tags" : [ "events" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "slug",
          "schema" : {
            "type" : "string"
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Event information",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Event"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Player" : {
        "type" : "object",
        "required" : [ "_id", "name", "slug", "nationality" ],
        "properties" : {
          "_id" : {
            "type" : "string",
            "example" : "660dc8c51e76999615f9e39f"
          },
          "name" : {
            "type" : "string",
            "example" : "Joshua Schmidt"
          },
          "slug" : {
            "type" : "string",
            "example" : "joshua-schmidt"
          },
          "nationality" : {
            "type" : "string",
            "example" : "DE"
          },
          "socials" : {
            "description" : "Object of links to the player's socials",
            "type" : "object",
            "properties" : {
              "youtube" : {
                "type" : "string",
                "example" : "JoshuaSchmidtYGO"
              },
              "twitch" : {
                "type" : "string",
                "example" : "joshuaschmidtygo"
              },
              "twitter" : {
                "type" : "string",
                "example" : "Gamebreak0r"
              }
            }
          }
        }
      },
      "Event" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "object",
            "properties" : {
              "name" : {
                "description" : "The name of the event type",
                "type" : "string",
                "example" : "YCS"
              },
              "slug" : {
                "description" : "The slug of the event type",
                "type" : "string",
                "example" : "ycs"
              }
            },
            "description" : "The type of the event"
          },
          "winner" : {
            "$ref" : "#/components/schemas/Winner"
          },
          "_id" : {
            "description" : "The id of the event",
            "type" : "string",
            "example" : "6620d3648d69215e8c9efe74"
          },
          "name" : {
            "description" : "The name of the event",
            "type" : "string",
            "example" : "YCS Bologna 2023"
          },
          "slug" : {
            "description" : "The slug of the event",
            "type" : "string",
            "example" : "bologna-2023"
          },
          "thumbnail" : {
            "description" : "The image identifier used to represent the event",
            "type" : "string",
            "example" : "bologna"
          },
          "topcut" : {
            "description" : "The top cut of the event",
            "type" : "integer",
            "example" : 64
          },
          "location" : {
            "description" : "The location of the event",
            "type" : "string",
            "example" : "Bologna, Italy"
          },
          "attendance" : {
            "description" : "The attendance of the event",
            "type" : "integer",
            "example" : 2488
          },
          "endDate" : {
            "description" : "The end date of the event",
            "type" : "string",
            "format" : "date-time",
            "example" : "2023-12-10T00:00:00.000Z"
          },
          "startDate" : {
            "description" : "The start date of the event",
            "type" : "string",
            "format" : "date-time",
            "example" : "2023-12-09T00:00:00.000Z"
          }
        }
      },
      "DeckType" : {
        "type" : "object",
        "required" : [ "_id", "name", "slug", "engines", "thumbnail" ],
        "properties" : {
          "_id" : {
            "description" : "The id of the deck type",
            "type" : "string",
            "example" : "6613201cdcb8246e3a5b59d1"
          },
          "name" : {
            "description" : "The name of the deck type",
            "type" : "string",
            "example" : "Runick Bystial"
          },
          "slug" : {
            "description" : "The slug for the deck type",
            "type" : "string",
            "example" : "runick-bystial"
          },
          "engines" : {
            "description" : "An array of engines used in the deck type",
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "example" : [ "Runick", "Bystial" ]
          },
          "thumbnail" : {
            "description" : "The passcode of the card used to represent the deck type",
            "type" : "string",
            "example" : 92107604
          }
        }
      },
...
      "Decklist" : {
        "type" : "object",
        "required" : [ "_id", "placement", "deckType", "event", "player" ],
        "properties" : {
          "_id" : {
            "description" : "The id of the decklist",
            "type" : "string",
            "example" : "6662f76ed85c2977d3561c3c"
          },
          "placement" : {
            "description" : "Placement of the decklist at the event",
            "type" : "integer",
            "example" : 1
          },
          "deckType" : {
            "$ref" : "#/components/schemas/DeckType"
          },
          "event" : {
            "$ref" : "#/components/schemas/Event"
          },
          "player" : {
            "$ref" : "#/components/schemas/Player"
          },
          "decklist" : {
            "description" : "The decklist itself",
            "type" : "object",
            "properties" : {
              "mainDeck" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Card"
                }
              },
              "extraDeck" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Card"
                }
              },
              "sideDeck" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Card"
                }
              }
            }
          }
        }
      },
      "Winner" : {
        "type" : "object",
        "required" : [ "deck", "player" ],
        "properties" : {
          "deck" : {
            "$ref" : "#/components/schemas/Decklist"
          },
          "player" : {
            "$ref" : "#/components/schemas/Player"
          }
        }
      },
      "Card" : {
        "type" : "object",
        "properties" : {
          "count" : {
            "description" : "Amount of copies",
            "type" : "number",
            "enum" : [ 1, 2, 3 ]
          },
          "card" : {
            "description" : "Id of the card",
            "type" : "string"
          }
        }
      }
    }
  }
}